--- docs/commands/teuthology-nuke.rst
@@ -1,4 +1,0 @@
-teuthology-nuke
-===============
-
-.. program-output:: teuthology-nuke --help
