--- roles/overrides.yaml
@@ -1,10 +1,0 @@
-nuke-on-error: true
-kernel:
-  branch: main
-overrides:
-  ceph:
-    branch: BRANCH_NAME
-    log-ignorelist:
-    - 'clocks not synchronized'
-tasks:
-- chef:
