--- scripts/test/test_nuke.py
@@ -1,5 +1,0 @@
-from script import Script
-
-
-class TestNuke(Script):
-    script_name = 'teuthology-nuke'
