--- teuthology/openstack/test/suites/nuke/nuke.yaml
@@ -1,8 +1,0 @@
-stop_worker: true
-nuke-on-error: true
-roles:
-- - client.0
-tasks:
-- exec:
-    client.0:
-      - exit 1
